+++ lib/src/media_viewer.dart
@@ -5,6 +5,7 @@ import 'models/media_viewer_config.dart';
 import 'widgets/image_viewer_widget.dart';
 import 'widgets/video_viewer_widget.dart';
 import 'widgets/youtube_viewer_widget.dart';
+import 'widgets/vimeo_viewer_widget.dart';
 import 'widgets/page_indicator_widget.dart';
 import 'widgets/navigation_arrows_widget.dart';
 import 'widgets/back_button_widget.dart';
@@ -167,6 +168,23 @@ class _MediaViewerState extends State<MediaViewer> {
                 onTap: _handleTap,
                 child: YouTubeViewerWidget(
                   key: ValueKey('youtube_$actualIndex'),
+                  item: item,
+                  config: widget.config,
+                  autoPlay: actualIndex == _currentIndex,
+                  onPlayingStateChanged: (isPlaying) {
+                    if (widget.config.hideArrowsWhenVideoPlays) {
+                      setState(() {
+                        _isVideoPlaying = isPlaying;
+                      });
+                    }
+                  },
+                ),
+              );
+            } else if (item.isVimeo) {
+              return GestureDetector(
+                onTap: _handleTap,
+                child: VimeoViewerWidget(
+                  key: ValueKey('vimeo_$actualIndex'),
                   item: item,
                   config: widget.config,
                   autoPlay: actualIndex == _currentIndex,